+++ R/RcppExports.R
@@ -5,6 +5,10 @@ DEVTRAN <- function(parin, inpar, parnames, init, cmtnames, capture, funs, data,
     .Call(`_mrgsolve_DEVTRAN`, parin, inpar, parnames, init, cmtnames, capture, funs, data, idata, OMEGA, SIGMA, envir)
 }
 
+MRGSIMQ <- function(parin, inpar, parnames, init, cmtnames, capture, funs, data, OMEGA, SIGMA, envir) {
+    .Call(`_mrgsolve_MRGSIMQ`, parin, inpar, parnames, init, cmtnames, capture, funs, data, OMEGA, SIGMA, envir)
+}
+
 MVGAUSS <- function(OMEGA_, n) {
     .Call(`_mrgsolve_MVGAUSS`, OMEGA_, n)
 }
@@ -31,9 +35,5 @@ TOUCH_FUNS <- function(lparam, linit, Neta, Neps, capture, funs, envir) {
 
 QUICKSIM <- function(parin, param, init, parnames, data, n, idata, req, capturei, funs, nre) {
     .Call(`_mrgsolve_QUICKSIM`, parin, param, init, parnames, data, n, idata, req, capturei, funs, nre)
-}
-
-SIMDATA <- function(parin, inpar, parnames, init, cmtnames, capture, funs, data, OMEGA, SIGMA, envir) {
-    .Call(`_mrgsolve_SIMDATA`, parin, inpar, parnames, init, cmtnames, capture, funs, data, OMEGA, SIGMA, envir)
 }